+++ libs/libc/string.h
@@ -244,15 +244,15 @@ INTDEF u64 LIBCCALL libc_strtou64_l(char const *__restrict nptr, char **__restri
 #define ERRNOSTR_NAME   offsetof(struct errnotext_entry,ete_name)
 #define ERRNOSTR_TEXT   offsetof(struct errnotext_entry,ete_text)
 struct errnotext_entry {
- u16       ete_name; /*< Offset into 'sed_strtab', to the name of the error. */
- u16       ete_text; /*< Offset into 'sed_strtab', to a string describing the error. */
+ u16 ete_name; /*< Offset into 'sed_strtab', to the name of the error. */
+ u16 ete_text; /*< Offset into 'sed_strtab', to a string describing the error. */
 };
 struct errnotext_data {
- uintptr_t              etd_version; /*< == STRERROR_VERSION (Strerror-data version; will always be backwards-compatible). */
- uintptr_t              etd_strtab;  /*< Offset from `strerror_data' to a string table. */
- uintptr_t              etd_enotab;  /*< Offset from `strerror_data' to a vector of `struct errnotext_entry'. */
- size_t                 etd_enocnt;  /*< Amount of entires in `etd_enotab'. */
- size_t                 etd_enoent;  /*< Size of a single entry within `etd_enotab'. */
+ u32 etd_version; /*< == STRERROR_VERSION (Strerror-data version; will always be backwards-compatible). */
+ s32 etd_strtab;  /*< Offset from `strerror_data' to a string table. */
+ s32 etd_enotab;  /*< Offset from `strerror_data' to a vector of `struct errnotext_entry'. */
+ u32 etd_enocnt;  /*< Amount of entires in `etd_enotab'. */
+ u32 etd_enoent;  /*< Size of a single entry within `etd_enotab'. */
 };
 
 #ifndef __KERNEL__
